lab3 nem vegleges

+++ projekt_szerver.py
@@ -12,6 +12,15 @@ import mongo
 def main():
     def create_folder():
         source_path = filedialog.askdirectory(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         path = os.path.join(source_path,input_variable)
         if(os.path.exists(path)):
             error_window=Tk()
@@ -27,12 +36,30 @@ def main():
 
     def drop_folder():
         source_path = filedialog.askdirectory(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         shutil.rmtree(source_path)
         seged=source_path.split("/")
         mongo.drop_database(seged[len(seged)-1])
 
     def create_table():
         source_path = filedialog.askdirectory(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         tabla=source_path+"\\"+tabla_nev+".json"
         if(os.path.exists(tabla)):
             error_window=Tk()
@@ -51,8 +78,16 @@ def main():
 
     def create_column():
         source_path = filedialog.askopenfilename(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         f=open(source_path,'a+')
-
         okes=0
         if(oszlop_index=='foreign-key'):
             referencia = filedialog.askopenfilename(title='Select reference')
@@ -68,7 +103,7 @@ def main():
             with open(referencia, 'r') as g:
                 data=json.load(g)
             for i in data:
-                if(i['column role']=='primary-key'):
+                if(i['column role']=='primary-key-unique' or i['column role']=='primary-key-not-unique'):
                     okes=1
                     break
                 else:
@@ -93,16 +128,32 @@ def main():
             f.write(",")
             f.write("\n")
 
-        if(oszlop_index=='foreign-key'):         
-            data={"column name":oszlop_nev,"column type":oszlop_tipus,"column role":oszlop_index,"references":referencia}
+        if(oszlop_index == 'foreign-key'):         
+            data = {"column name":oszlop_nev,"column type":oszlop_tipus,"column role":oszlop_index,"references":referencia}
+            fk = oszlop_nev+' '+oszlop_index + ' ' + referencia
         else:
-            data={"column name":oszlop_nev,"column type":oszlop_tipus,"column role":oszlop_index}
+            data = {"column name":oszlop_nev,"column type":oszlop_tipus,"column role":oszlop_index}
+            pk=oszlop_nev+' '+oszlop_index
         json.dump(data,f,indent=4)
         f.write("\n]")
         f.close()
+        seged = source_path.split("/")  
+        if(oszlop_index == 'primary-key-unique' or oszlop_index == 'primary-key-not-unique'):
+            mongo.add_column(pk, seged[len(seged)-2], seged[len(seged)-1])
+        elif(oszlop_index=='foreign-key'):
+            mongo.add_column(fk, seged[len(seged)-2], seged[len(seged)-1])
 
     def drop_table():
         source_path = filedialog.askopenfilename(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         os.remove(source_path)
         seged=source_path.split("/")
         collection_nev=seged[len(seged)-1].split(".")
@@ -110,10 +161,28 @@ def main():
 
     def insert():
         source_path = filedialog.askopenfilename(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         client.send(source_path.encode('utf-8'))
     
     def delete():
         source_path=filedialog.askopenfilename(title='Select Title')
+        if(source_path==""):
+            error_window=Tk()
+            lbl_error = Label(error_window,text="No cancel pls", fg='red', font=('Times New Roman', 16), bg="lightblue")
+            lbl_error.place(x=10,y=50)
+            error_window.title('Error')
+            error_window.geometry("400x100+10+20")
+            error_window.lift()
+            error_window.mainloop()
+            return
         client.send(source_path.encode('utf-8'))
 
     def check_insert():
@@ -211,9 +280,9 @@ def main():
             tabla_nev=valtozok[1]
             create_table()
         elif(valtozok[0]=='create_column'):
-            oszlop_nev=valtozok[1]
-            oszlop_tipus=valtozok[2]
-            oszlop_index=valtozok[3]
+            oszlop_nev = valtozok[1]
+            oszlop_tipus = valtozok[2]
+            oszlop_index = valtozok[3]
             create_column()
         elif(valtozok[0]=='drop_table'):
             drop_table()
